requirements.txt
aiogram==2.10.1
python-dotenv==0.14.0
